+++ src/main/kotlin/dev/realmkit/game/domain/aliases/Aliases.kt
@@ -20,16 +20,16 @@
 
 package dev.realmkit.game.domain.aliases
 
+import dev.realmkit.game.core.document.CurrentMax
 import dev.realmkit.game.domain.battle.action.BattleAction
 import dev.realmkit.game.domain.battle.action.BattleActionAttack
-import dev.realmkit.game.domain.stat.document.StatValue
 import dev.realmkit.game.domain.target.document.Target
 
 /**
- * # [StatValueDouble]
- * the `stat value` [Double] alias
+ * # [CurrentMaxDouble]
+ * the `current max` [Double] alias
  */
-typealias StatValueDouble = StatValue<Double>
+typealias CurrentMaxDouble = CurrentMax<Double>
 
 /**
  * # [AttackBlock]