+++ src/test/kotlin/dev/realmkit/game/domain/battle/service/BattleServiceTest.kt
@@ -45,8 +45,7 @@ class BattleServiceTest(
                     Player.fixture,
                     Player.fixture,
                 ) { player, enemy ->
-                    player.stat.base.shield.current = 100.0
-                    player.stat.base.power = 100.0
+                    player.stat.base.attack = 100.0
                     player.stat.base.defense = 100.0
                     player.stat.base.speed = 1.0
                     player.stat.multiplier.critical = 2.0
@@ -64,8 +63,7 @@ class BattleServiceTest(
                     Player.fixture,
                     Player.fixture,
                 ) { player, enemy1, enemy2 ->
-                    player.stat.base.shield.current = 100.0
-                    player.stat.base.power = 100.0
+                    player.stat.base.attack = 100.0
                     player.stat.base.defense = 100.0
                     player.stat.base.speed = 1.0
                     player.stat.multiplier.critical = 1.0
@@ -84,9 +82,7 @@ class BattleServiceTest(
                     Player.fixture,
                     Player.fixture,
                 ) { player1, player2, enemy ->
-                    player1.stat.base.shield.current = 100.0
-                    player1.stat.base.power = 100.0
-                    player1.stat.base.defense = 100.0
+                    player1.stat.base.attack = 100.0
                     player1.stat.base.speed = 1.0
                     player1.stat.multiplier.critical = 1.0
                     player2.stat.base.defense = 100.0
@@ -108,12 +104,9 @@ class BattleServiceTest(
                     Player.fixture,
                     Player.fixture,
                 ) { player1, player2, enemy1, enemy2 ->
-                    player1.stat.base.shield.current = 100.0
-                    player1.stat.base.power = 100.0
-                    player1.stat.base.defense = 100.0
+                    player1.stat.base.attack = 100.0
                     player1.stat.base.speed = 2.0
                     player1.stat.multiplier.critical = 1.0
-                    player2.stat.base.defense = 100.0
                     player2.stat.base.speed = 1.0
                     player2.stat.multiplier.critical = 1.0
 
@@ -139,12 +132,9 @@ class BattleServiceTest(
                     player1.stat.base.speed = 1.0
                     player2.stat.base.speed = 0.5
 
-                    enemy1.stat.base.shield.current = 100.0
-                    enemy1.stat.base.power = 100.0
-                    enemy1.stat.base.defense = 100.0
+                    enemy1.stat.base.attack = 100.0
                     enemy1.stat.base.speed = 2.0
                     enemy1.stat.multiplier.critical = 1.0
-                    enemy2.stat.base.defense = 100.0
                     enemy2.stat.base.speed = 1.0
                     enemy2.stat.multiplier.critical = 1.0
 
@@ -162,8 +152,8 @@ class BattleServiceTest(
                     Player.fixture,
                     Player.fixture,
                 ) { player, enemy ->
-                    player.stat.base.power = 0.0
-                    enemy.stat.base.power = 0.0
+                    player.stat.base.attack = 0.0
+                    enemy.stat.base.attack = 0.0
 
                     battleService.battle { player against enemy }